requirements.txt
### `requirements.txt`

torch==2.2.0
torchvision==0.17.0+cu121
pytorch-lightning==2.2.2
timm    #==0.6.13
mlflow  #==2.12.2
ray[tune]==2.20.0
torchmetrics    #==0.8.2
pyyaml  #==6.0.1
Pillow  #==9.4.0
imgaug  #==0.4.0
einops  #==0.7.0
matplotlib  #==3.7.2
tensorboard #==2.8.0
numpy   #==1.23.4